moved subscription package into core, since it's the only thing that uses it. Update subscription tests
